% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inline-text.R
\name{apa.rcorr.inline}
\alias{apa.rcorr.inline}
\title{Report correlation between two variables}
\usage{
apa.rcorr.inline(rcorr, var1, var2, type = "pearson")
}
\arguments{
\item{var1}{Name of first variable of interest (string/quoted object)}

\item{var2}{Name of second variable of interest (string/quoted object)}

\item{type}{The method of correlation used to generate the \code{rcorr}
  matrix}

\item{x}{An \code{rcorr} matrix}
}
\description{
Displays the correlation between two variables within an
\code{\link[Hmisc]{rcorr}} matrix table according to APA 6 guidelines.
}
\examples{
iris.rcorr <- Hmisc::rcorr(as.matrix(iris[, 1:4]))
apa.rcorr.inline(iris.rcorr, 'Sepal.Length', 'Sepal.Width')
}

