% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apa.R
\name{apa}
\alias{apa}
\title{Helper function for generating APA 6-styled output from R objects}
\usage{
apa(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Arguments to be passed on to the applicable \code{rapa} function.}
}
\description{
\code{apa} takes an R object, finds an applicable \code{rapa} function, and
returns a string of APA 6-styled text, if an appropriate function is
available; otherwise, returns an empty string.
}
\examples{
library(Hmisc)
library(xtable)

iris.rcorr <- rcorr(as.matrix(iris[, 1:4]))
iris.rcorr.pp <- rcorr.pp(iris.rcorr)
iris.xtable <- xtable(iris.rcorr.pp)

# Return correlation
apa(iris.rcorr, 'Sepal.Length', 'Sepal.Width')

# Return xtable
apa(iris.xtable)
}

