% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tables.R
\name{apa.stargazer}
\alias{apa.stargazer}
\title{Create APA 6-styled regression tables}
\usage{
apa.stargazer(...)
}
\arguments{
\item{...}{Parameters passed on to the \code{stargazer} function}
}
\description{
Function that customizes the \code{\link[stargazer]{stargazer}} in order that
it produces regression model tables that conform to APA 6 guidelines.
}
\details{
Customized parametres, which should not be otherwise set, include:
\code{style},
\code{column.labels},
\code{notes.align},
\code{notes.label},
\code{keep.stat},
\code{no.space},
and
\code{header}.
No new parametres are added, so refer to the \code{stargazer} documentation
for complete usage information.

Note: At this time, adjusted R-squared is labelled \code{Adjusted R$^{2}$}
rather than \code{\\delta R$^{2}$} in this function's output. Until someone is
able to add an APA style to the \code{stargazer} package, this output is as
close to APA 6-style as this function will get.
}
\examples{
library(stargazer)

iris.lm.1 <- lm(Sepal.Length ~ Sepal.Width, data = iris)
iris.lm.2 <- update(iris.lm.1, .~. + Petal.Length + Petal.Width)

apa.stargazer(iris.lm.1, iris.lm.2,
              covariate.labels = c("1. SW", "2. PL", "2. PW"),
              notes = "SW = Sepal width, PL = Petal length, PW = Petal width"
)
}
\seealso{
\code{\link[stargazer]{stargazer}}
}

