% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tables.R
\name{apa.xtable}
\alias{apa.xtable}
\title{Create APA 6 export tables}
\usage{
apa.xtable(xtable, note = NULL, wrap.text = FALSE, ...)
}
\arguments{
\item{note}{Character vector of length 1 containing the table's notes.}
}
\description{
Function that take an \code{\link[xtable]{xtable}} object and prints a LaTeX
table in APA 6 format. It uses the same arguments as the original
\code{\link[xtable]{xtable.print}} function as well as an additional
argument for creating table notes.
}
\seealso{
\code{\link[xtable]{xtable}}, \code{\link[xtable]{print.xtable}}
}

