% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rapa-package.R
\docType{package}
\name{rapa-package}
\alias{rapa-package}
\title{Generate APA 6-styled output from R objects.}
\description{
\code{rapa} allows you to dynamically inject R objects (especially tables)
into your manuscript and does its best to style them in accordance with
APA 6 guidelines.
}
\details{
Provides a series of helper functions that return APA-styled text in
\href{http://www.latex-project.org}{LaTeX} or
\href{http://pandoc.org/index.html}{pandoc-style markdown} format.
Especially handy when used with
\href{http://rmarkdown.rstudio.com}{RMarkdown}.

Please contact the author if additional helper functions are desired.
}

