% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inline-text.R
\name{apa.p.value}
\alias{apa.p.value}
\title{Report a p-value}
\usage{
apa.p.value(p, pretty.print = TRUE)
}
\arguments{
\item{p}{A number representing a p-value (must be less than 1).}

\item{pretty.print}{An optional dichotomous indicator for whether to prepend
the p-value abbreviation and appropriate sign to the returned number (see
details).}
}
\description{
Formats and reports a p-value according to APA 6 guidelines. Returns a
string.
}
\details{
This function is primarily intended to be used by other helper functions, so
by default p-values are pretty-printed (e.g., "_p_ = .152") instead of simply
returning the correctly rounded p-value (e.g., ".152")
}

