% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tables.R
\name{apa.xtable}
\alias{apa.xtable}
\title{Create APA 6 export tables}
\usage{
apa.xtable(xtable, note = NULL, wrap.text = FALSE, ...)
}
\arguments{
\item{xtable}{An \code{xtable} object.}

\item{note}{Character vector of length 1 containing the table's notes.}

\item{wrap.text}{Quick helper that uses the 'tabularx' table environment and
allows for long notes.}

\item{...}{Arguments passed to \code{print.xtable}.}
}
\description{
Function that take an \code{\link[xtable]{xtable}} object and prints a LaTeX
table in APA 6 format. It uses the same arguments as the original
\code{\link[xtable]{print.xtable}} function as well as an additional
argument for creating table notes.
}
\examples{
library(Hmisc)
library(xtable)

iris.rcorr <- rcorr(as.matrix(iris[, 1:4]))
iris.rcorr.pp <- rcorr.pp(iris.rcorr)
iris.xtable <- xtable(iris.rcorr.pp)

apa.xtable(iris.xtable)
}
\seealso{
\code{\link[xtable]{xtable}}, \code{\link[xtable]{print.xtable}}
}

